Reference

Contents

Reference

Contents

Index

Index

NCL.NCLModel

NCL.NCLModel — Type

NCLModel(nlp)

Subtype of AbstractNLPModel designed to represent an NCL subproblem. A general problem of the form

minimize   f(x)
over       x
subject to lvar ≤ x ≤ uvar
           lcon ≤ c(x) ≤ ucon

is transformed into

minimize   f(x) + λ'r + ρ ‖r‖²
over       x, r
subject to lvar ≤ x ≤ uvar
           lcon ≤ c(x) + r ≤ ucon

where λ is a vector of Lagrange multiplier estimates and ρ > 0 is a penalty parameter.

Input arguments

nlp::AbstractNLPModel the original problem

Keyword arguments

resid::Float64 the initial residual value (default 0)

resid_linear::Bool whether or not residuals are added to linear constraints

ρ::Float64 initial penalty parameter

y::AbstractVector{Float64} initial Lagrange multiplier estimates

Return value

ncl::NCLModel the transformed model.