Reference

Contents

Reference

Contents

Index

Index

NCL.AbstractNCLSubSolver

NCL.NCLModel

NCL.IpoptNCLSubSolver

NCL.KnitroNCLSubSolver

NCL.MadNLPNCLSubSolver

NCL.NCLSolve

NCL.add_to_multipliers!

NCL.get_multipliers

NCL.get_nlp

NCL.get_nr

NCL.get_nx

NCL.get_penalty_parameter

NCL.get_resid_linear

NCL.set_penalty_parameter!

NCL.AbstractNCLSubSolver — Type

AbstractNCLSubSolver

Abstract type from which NCL subproblem solvers derive. A subproblem solver must be a callable struct and should return an AbstractExecutionStats instance (that can be stored inside the solver and updated for efficiency).

NCL subproblem solvers conform to the following API:

Constructor mandatory arguments

ncl_model::NCLModel: the NCL model to be solved as a subproblem.

Constructor keyword arguments

dfeas_abs_tol::eltype(ncl_model): the absolute tolerance on dual feasibility;

pfeas_abs_tol::eltype(ncl_model): the absolute tolerance on primal feasibility;

compl_abs_tol::eltype(ncl_model): the absolute tolerance on complementarity.

Additional keyword arguments are passed to the inner solver constructor. Such arguments can be used, e.g., with the KnitroNCLSubSolver to set linear_api = false in case the model type used does not support the linear API (e.g., an AmplModel).

Solver arguments

ncl_model::NCLModel: the NCL model to be solved as a subproblem;

outer_iter::Int: the outer iteration counter;

rel_tol::eltype(ncl_model): the relative stopping tolerance.

Solver keyword arguments

x0::AbstractVector = get_x0(ncl_model): the initial guess.

All other keyword arguments will be passed directly to the underlying solver.

NCL.NCLModel — Type

NCLModel(nlp)

Subtype of AbstractNLPModel designed to represent an NCL subproblem. A general problem of the form

minimize   f(x)
over       x
subject to lvar ≤ x ≤ uvar
           lcon ≤ c(x) ≤ ucon

is transformed into

minimize   f(x) + λ'r + 1/2 ρ ‖r‖²
over       x, r
subject to lvar ≤ x ≤ uvar
           lcon ≤ c(x) + r ≤ ucon

where λ is a vector of Lagrange multiplier estimates and ρ > 0 is a penalty parameter.

Input arguments

nlp::AbstractNLPModel the original problem

Keyword arguments

resid::Float64 the initial residual value (default 0)

resid_linear::Bool whether or not residuals are added to linear constraints

ρ::Float64 initial penalty parameter

y::AbstractVector{Float64} initial Lagrange multiplier estimates

Return value

ncl::NCLModel the transformed model.

NCL.IpoptNCLSubSolver — Function

AbstractNCLSubSolver

Abstract type from which NCL subproblem solvers derive. A subproblem solver must be a callable struct and should return an AbstractExecutionStats instance (that can be stored inside the solver and updated for efficiency).

NCL subproblem solvers conform to the following API:

Constructor mandatory arguments

ncl_model::NCLModel: the NCL model to be solved as a subproblem.

Constructor keyword arguments

dfeas_abs_tol::eltype(ncl_model): the absolute tolerance on dual feasibility;

pfeas_abs_tol::eltype(ncl_model): the absolute tolerance on primal feasibility;

compl_abs_tol::eltype(ncl_model): the absolute tolerance on complementarity.

Additional keyword arguments are passed to the inner solver constructor. Such arguments can be used, e.g., with the KnitroNCLSubSolver to set linear_api = false in case the model type used does not support the linear API (e.g., an AmplModel).

Solver arguments

ncl_model::NCLModel: the NCL model to be solved as a subproblem;

outer_iter::Int: the outer iteration counter;

rel_tol::eltype(ncl_model): the relative stopping tolerance.

Solver keyword arguments

x0::AbstractVector = get_x0(ncl_model): the initial guess.

All other keyword arguments will be passed directly to the underlying solver.

NCL.KnitroNCLSubSolver — Function

AbstractNCLSubSolver

Abstract type from which NCL subproblem solvers derive. A subproblem solver must be a callable struct and should return an AbstractExecutionStats instance (that can be stored inside the solver and updated for efficiency).

NCL subproblem solvers conform to the following API:

Constructor mandatory arguments

ncl_model::NCLModel: the NCL model to be solved as a subproblem.

Constructor keyword arguments

dfeas_abs_tol::eltype(ncl_model): the absolute tolerance on dual feasibility;

pfeas_abs_tol::eltype(ncl_model): the absolute tolerance on primal feasibility;

compl_abs_tol::eltype(ncl_model): the absolute tolerance on complementarity.

Additional keyword arguments are passed to the inner solver constructor. Such arguments can be used, e.g., with the KnitroNCLSubSolver to set linear_api = false in case the model type used does not support the linear API (e.g., an AmplModel).

Solver arguments

ncl_model::NCLModel: the NCL model to be solved as a subproblem;

outer_iter::Int: the outer iteration counter;

rel_tol::eltype(ncl_model): the relative stopping tolerance.

Solver keyword arguments

x0::AbstractVector = get_x0(ncl_model): the initial guess.

All other keyword arguments will be passed directly to the underlying solver.

NCL.MadNLPNCLSubSolver — Function

AbstractNCLSubSolver

Abstract type from which NCL subproblem solvers derive. A subproblem solver must be a callable struct and should return an AbstractExecutionStats instance (that can be stored inside the solver and updated for efficiency).

NCL subproblem solvers conform to the following API:

Constructor mandatory arguments

ncl_model::NCLModel: the NCL model to be solved as a subproblem.

Constructor keyword arguments

dfeas_abs_tol::eltype(ncl_model): the absolute tolerance on dual feasibility;

pfeas_abs_tol::eltype(ncl_model): the absolute tolerance on primal feasibility;

compl_abs_tol::eltype(ncl_model): the absolute tolerance on complementarity.

Additional keyword arguments are passed to the inner solver constructor. Such arguments can be used, e.g., with the KnitroNCLSubSolver to set linear_api = false in case the model type used does not support the linear API (e.g., an AmplModel).

Solver arguments

ncl_model::NCLModel: the NCL model to be solved as a subproblem;

outer_iter::Int: the outer iteration counter;

rel_tol::eltype(ncl_model): the relative stopping tolerance.

Solver keyword arguments

x0::AbstractVector = get_x0(ncl_model): the initial guess.

All other keyword arguments will be passed directly to the underlying solver.

NCL.NCLSolve — Method

NCLSolve(ncl::NCLModel; kwargs...)
NCLSolve(nlp::AbstractNLPModel; kwargs...)

Solve problem ncl using Algorithm NCL. If nlp is not already an NCLModel, it will be converted to one, and the default subproblem solver will be used. If nlp does not have constraints other than bound constraints, an error will be thrown. In that case, nlp should be passed directly to an unconstrained or bound-constrained solver. It could also be passed directly to one of the NCL subproblem solvers that are currently supported.

The solver stops under the following circumstances:

The norm of the residual variables falls below feas_tol and the dual feasibility residual, as reported by the subproblem solver, falls below opt_tol;

the maximum number of outer iterations max_iter_NCL is attained;

the problem is found to be (locally) infeasible.

Keyword arguments

opt_tol::Float64 = 1.0e-6: dual feasibility stopping tolerance;

feas_tol::Float64 = 1.0e-6: primal feasibility stopping tolerance;

max_iter_NCL::Int = 20: maximum number of outer iterations;

subsolver::AbstractNCLSubSolver = IpoptNCLSubSolver(ncl): subproblem solver;

verbose::Bool = true: whether to display progress information at each iteration.

Additional keyword arguments will be passed directly to the subproblem solver.

NCL.add_to_multipliers! — Method

 add_to_multipliers!(ncl::NCLModel, α, v)

Update the Lagrange multiplier estimates by y ← y + α v, where α is a scalar and v is a vector of the same length as y.

NCL.get_multipliers — Method

get_multipliers(ncl::NCLModel)

Return the current vector of Lagrange multiplier estimates y.

NCL.get_nlp — Method

get_nlp(ncl::NCLModel)

Return the underlying model of ncl.

NCL.get_nr — Method

 get_nr(ncl::NCLModel)

Return the number of residuals added in the NCL problem (equal to the number of constraints in the original problem if resid_linear is true, and to the number of nonlinear constraints otherwise).

NCL.get_nx — Method

 get_nx(ncl::NCLModel)

Return the number of variables in the original problem (not including residuals).

NCL.get_penalty_parameter — Method

get_penalty_parameter(ncl::NCLModel)

Return the current value of the penalty parameter ρ.

NCL.get_resid_linear — Method

 get_resid_linear(ncl::NCLModel)

Return whether or not the residuals are added to linear constraints (if true, residuals are added to linear constraints; if false, residuals are added to nonlinear constraints only).

NCL.set_penalty_parameter! — Method

set_penalty_parameter!(ncl::NCLModel, ρ)

Set the penalty parameter to ρ > 0.